Add backface culling to the way footage geo is baked and rendered.
Without this you can get artifacts in some situations due to the fact
that the same bake texture is mapped to both sides of a mesh face.
Another way to address this would be to map a different texture to
each side, but it's not clear how to make baking understand that.